Create Arduino PID code
Arduino Code for a PID control loop for DC motor integrated with quadrature encoder. Values Kp (proportional), Ki (integral) and Kd (derivative) must be tuned to achieve a desire and smooth angular position for the robot. It's recommended to derive the plant system for the DC motor and load of the robotic linkage. Then graph a PID step response graph using MATLAB and properly tune the Ki, Kd, and Kp to achieve a low overshoot, a low rise-time and settling time, and no steady-state error,