MATLAB Script for PID Simulation Program
Learn to create a PID simulation program for a DC motor linked to a timing belt and gears, to drive the linkages of a SCARA robot. Use the step response graphs to analyze how it fluctuates base on different values of Kp, Ki and Kd. These coefficients can be transposed to an Arduino code and be tested on the actual hardware.